requirements.txt
numpy==1.26.4
obsws-python==1.7.0
openai==1.30.5
pytchat==0.5.5
python-dotenv==1.0.1
requests==2.32.3
sounddevice==0.4.7
soundfile==0.12.1
